move makred_timer a special implementation in nvtx
