+++ verl/trainer/ppo/ray_trainer.py
@@ -51,7 +51,7 @@ from verl.trainer.ppo.metric_utils import (
 )
 from verl.trainer.ppo.reward import compute_reward, compute_reward_async
 from verl.utils.checkpoint.checkpoint_manager import BaseCheckpointManager, find_latest_ckpt_path
-from verl.utils.debug.performance import marked_timer
+from verl.utils.debug import marked_timer
 from verl.utils.metric import (
     reduce_metrics,
 )
