+++ verl/utils/debug/nvtx_profile.py
@@ -14,7 +14,8 @@
 # limitations under the License.
 
 import functools
-from typing import Callable, Optional
+from contextlib import contextmanager
+from typing import Callable, Dict, Optional
 
 import nvtx
 import torch
@@ -67,6 +68,28 @@ def mark_annotate(message: Optional[str] = None, color: Optional[str] = None, do
         return nvtx.annotate(profile_message, color=color, domain=domain, category=category)(func)
 
     return decorator
+
+
+@contextmanager
+def marked_timer(name: str, timing_raw: Dict[str, float], color: str = None):
+    """Context manager for timing with NVTX markers.
+
+    This utility function measures the execution time of code within its context,
+    accumulates the timing information, and adds NVTX markers for profiling.
+
+    Args:
+        name (str): The name/identifier for this timing measurement.
+        timing_raw (Dict[str, float]): Dictionary to store timing information.
+        color (Optional[str]): Color for the NVTX marker. Defaults to None.
+
+    Yields:
+        None: This is a context manager that yields control back to the code block.
+    """
+    mark_range = mark_start_range(message=name, color=color)
+    from .performance import _timer
+
+    yield from _timer(name, timing_raw)
+    mark_end_range(mark_range)
 
 
 class NsightSystemsProfiler(WorkerProfiler):
